+++ lib/bart/bart_appbar.dart
@@ -26,15 +26,18 @@ class BartAppbarAction extends Action<AppBarBuildIntent> {
 
 mixin AppBarNotifier {
   void updateAppBar(BuildContext context, PreferredSizeWidget? appBar) {
-    _runWhenReady(context, () => Actions.invoke(context, AppBarBuildIntent(appBar)));
+    _runWhenReady(
+        context, () => Actions.invoke(context, AppBarBuildIntent(appBar)));
   }
 
   void showAppBar(BuildContext context) {
-    _runWhenReady(context, () => Actions.invoke(context, AppBarAnimationIntent.show()));
+    _runWhenReady(
+        context, () => Actions.invoke(context, AppBarAnimationIntent.show()));
   }
 
   void hideAppBar(BuildContext context) {
-    _runWhenReady(context, () => Actions.invoke(context, AppBarAnimationIntent.hide()));
+    _runWhenReady(
+        context, () => Actions.invoke(context, AppBarAnimationIntent.hide()));
   }
 
   _runWhenReady(BuildContext context, Function onReady) {