+++ lib/bart/router_delegate.dart
@@ -14,16 +14,22 @@ class MenuRouter extends InheritedWidget {
     required BartRouteBuilder routesBuilder,
     List<NavigatorObserver>? navigatorObservers,
     required this.child,
-  })  : routerDelegate = MenuRouterDelegate(routesBuilder.call(), initialRoute, navigatorObservers),
+  })  : routerDelegate = MenuRouterDelegate(
+            routesBuilder.call(), initialRoute, navigatorObservers),
         super(key: key, child: child);
 
-  static MenuRouter of(BuildContext context) => context.dependOnInheritedWidgetOfExactType<MenuRouter>()!;
+  static MenuRouter of(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<MenuRouter>()!;
 
   @override
   bool updateShouldNotify(MenuRouter oldWidget) => true;
 }
 
-class MenuRouterDelegate extends RouterDelegate<MenuRoutePath> with ChangeNotifier, PopNavigatorRouterDelegateMixin<MenuRoutePath>, AppBarNotifier {
+class MenuRouterDelegate extends RouterDelegate<MenuRoutePath>
+    with
+        ChangeNotifier,
+        PopNavigatorRouterDelegateMixin<MenuRoutePath>,
+        AppBarNotifier {
   final List<BartMenuRoute> routes;
   final String? initialRoute;
   final List<NavigatorObserver>? navigatorObservers;
@@ -67,13 +73,15 @@ class MenuRouterDelegate extends RouterDelegate<MenuRoutePath> with ChangeNotifi
               pageBuilder: (context, __, ___) {
                 if (searchedRoute.cache) {
                   if (!pageCache.containsKey(searchedRoute.path)) {
-                    pageCache[searchedRoute.path] = searchedRoute.pageBuilder(context);
+                    pageCache[searchedRoute.path] =
+                        searchedRoute.pageBuilder(context);
                   }
                   return pageCache[searchedRoute.path]!;
                 }
                 return searchedRoute.pageBuilder(context);
               },
-              transitionsBuilder: (_, a, __, c) => FadeTransition(opacity: a, child: c));
+              transitionsBuilder: (_, a, __, c) =>
+                  FadeTransition(opacity: a, child: c));
           if (navigatorObservers != null) {
             navigatorObservers!.forEach((element) {
               element.didPush(pageRoute, null);