+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## [0.0.3] - Cache and Appbar animation
+* show or hide appbar with animation 
+* preserve page in cache (see readme)
+
 ## [0.0.1] - First release
 * first release